backend/requirements.txt
Flask
pymongo
python-dotenv
pandas
numpy
seaborn
pymongo[srv]
Flask-PyMongo
pythondns
flask-cors
nba_api